third_party/blink/web_tests/external/wpt/css/cssom-view/MediaQueryList-extends-EventTarget-interop-expected.txt

This is a testharness.js-based test.
PASS dispatchEvent triggers listener added with addListener
PASS listener added with addListener and addEventListener is called once
FAIL listener added with addListener and addEventListener (capture) is called twice assert_equals: triggerMQLEvent expected 2 but got 0
PASS removeEventListener removes listener added with addListener
FAIL removeEventListener (capture) doesn't remove listener added with addListener assert_equals: triggerMQLEvent expected 1 but got 0
PASS removeListener removes listener added with addEventListener
FAIL removeListener doesn't remove listener added with addEventListener (capture) assert_equals: triggerMQLEvent expected 1 but got 0
FAIL listeners are called in order they were added, ignoring capture parameter assert_array_equals: triggerMQLEvent lengths differ, expected 2 got 0
Harness: the test ran to completion.
